requirements.txt
kagglehub
pandas
python-dotenv
numpy
mne
pywavelets
scikit-learn
torch
transformers
typeguard
scipy
sortedcontainers
tensorflow-metal==1.1.0
tensorflow-macos==2.15.0